requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
aiofiles
pydantic
pillow
pytesseract
celery[redis]
dotenv
asyncpg
python-multipart
alembic
psycopg2-binary
django
requests
djangorestframework
prometheus_client
pyjwt
djangorestframework_simplejwt
djoser
loguru
pytz